resolved bugs

+++ gps_module/gps.cpp
@@ -129,7 +129,7 @@ UbxMessage Gps::readUbxMessage() {
       int8_t sync1_to_compare = i2c_smbus_read_byte_data(i2c_fd, DATA_STREAM_REGISTER);
       int8_t sync2_to_compare = i2c_smbus_read_byte_data(i2c_fd, DATA_STREAM_REGISTER);
 
-      printf("Sync1: 0x%x Sync2: 0x%x\n", sync1_to_compare, sync2_to_compare);
+      printf("Sync1: 0x%x Sync2: 0x%x\n", (uint8_t) sync1_to_compare, sync2_to_compare);
 
     if (sync1_to_compare == SYNC_CHAR_1 && sync2_to_compare == SYNC_CHAR_2) {
         for (int i = 0; i < messageLength; i++) {
